Move compile methods to irgen module. So far, the LuaParseTree structure was responsible for compiling Lua to bytecode. This doesn't really make sense, because a compiler is usually separated from the parse tree. The compiler usually works on a parse tree. This commit introduces a new module, irgen, in which the LuaToBytecode structure was added. This new structure takes a parse tree as input and produces bytecode for the lua vm. In the future, the frontend will translate the parse tree into a higher level IR, rather than into bytecode directly. This structure will allow us to add a register allocation algorithm, and also help with simple optimization such as constant propagation, and constant folding.
