+++ luacompiler/src/lib/irgen/constants_map.rs
