+++ luacompiler/src/lib/irgen/register_map.rs
